backend/requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7

# Google APIs and Authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI and External APIs
openai==1.3.7
requests==2.31.0

# Security and Encryption
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing Dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1
freezegun==1.2.2
